refactored the project to use the new evaluation method
+ removed the old method
