+++ lib/prompt-templates/index.ts
@@ -1,3 +1,1 @@
-export * from "./create-circuit-board1"
-export * from "./create-circuit-board2"
 export * from "./create-local-circuit-prompt"
