--- tests/board-samples/sample1.ts
@@ -1,1 +1,0 @@
-export const boardSample1 = "an led with an 0402 footprint"
