--- tests/board-samples/sample2.ts
@@ -1,1 +1,0 @@
-export const boardSample2 = "a blinking led connected to a 555 timer"
