--- tests/board-samples/sample3.ts
@@ -1,2 +1,0 @@
-export const boardSample3 =
-  "a push button that when pressed will turn on a light"
