--- tests/board-samples/sample4.ts
@@ -1,1 +1,0 @@
-export const boardSample4 = "a switch that when flipped spins a motor"
