--- tests/module-samples/sample1-na555.ts
@@ -1,1 +1,0 @@
-export const moduleSample1 = "the NA555 timer"
