--- tests/prompts/create-circuit-board1/create-circuit-board1-sample1.test.ts
@@ -1,22 +1,0 @@
-import { expect, test } from "bun:test"
-import { runInitialPrompt } from "lib/code-runner/run-prompt"
-import { createCircuitBoard1Template } from "lib/prompt-templates/create-circuit-board1"
-import { boardSample1 } from "tests/board-samples/sample1"
-
-test("create-circuit-board1-prompt1", async () => {
-  const systemPrompt = createCircuitBoard1Template({ currentCode: "" })
-
-  const { success, circuit } = await runInitialPrompt(
-    { systemPrompt, userPrompt: boardSample1 },
-    {
-      model: "gpt-4o-mini",
-      outputType: "board",
-    },
-  )
-
-  expect(success).toBe(true)
-
-  const led = circuit?.selectOne("led")
-
-  expect(led).toBeDefined()
-})
